[not verified] [not verified] podcast-player: finish renaming utils
